Move config test to pan de agua.
